requirements.txt
flask
flask-sqlalchemy
flask-login
Flask-Migrate
# For PostgreSQL
psycopg2-binary
python-dotenv